requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pdfplumber==0.11.4
openai==1.58.1
python-dotenv==1.0.1
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
